Adding account operations in auth
