+++ horusec-account/internal/handlers/companies/companies_test.go
@@ -19,7 +19,6 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
-	"github.com/ZupIT/horusec/development-kit/pkg/databases/relational/repository/cache"
 	"github.com/ZupIT/horusec/development-kit/pkg/entities/auth"
 	authEnums "github.com/ZupIT/horusec/development-kit/pkg/enums/auth"
 	"github.com/ZupIT/horusec/horusec-account/config/app"
@@ -60,7 +59,6 @@ func TestCreateCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockTx := &relational.MockWrite{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		company := &accountEntities.Company{
 			Name: "test",
@@ -73,7 +71,7 @@ func TestCreateCompany(t *testing.T) {
 
 		mockWrite.On("StartTransaction").Return(mockTx)
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 		body, _ := json.Marshal(company)
 		r, _ := http.NewRequest(http.MethodPost, "api/companies", bytes.NewReader(body))
 
@@ -96,7 +94,6 @@ func TestCreateCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockTx := &relational.MockWrite{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 		appConfig := &app.Config{}
 
 		company := &accountEntities.CompanyApplicationAdmin{
@@ -116,7 +113,7 @@ func TestCreateCompany(t *testing.T) {
 			AccountID: uuid.New(),
 		}))
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, appConfig)
+		handler := NewHandler(mockWrite, mockRead, brokerMock, appConfig)
 		body, _ := json.Marshal(company)
 		r, _ := http.NewRequest(http.MethodPost, "api/companies", bytes.NewReader(body))
 
@@ -141,7 +138,6 @@ func TestCreateCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockTx := &relational.MockWrite{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 		appConfig := &app.Config{}
 
 		company := &accountEntities.CompanyApplicationAdmin{
@@ -161,7 +157,7 @@ func TestCreateCompany(t *testing.T) {
 			AccountID: uuid.New(),
 		}))
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, appConfig)
+		handler := NewHandler(mockWrite, mockRead, brokerMock, appConfig)
 		body, _ := json.Marshal(company)
 		r, _ := http.NewRequest(http.MethodPost, "api/companies", bytes.NewReader(body))
 
@@ -183,7 +179,6 @@ func TestCreateCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockTx := &relational.MockWrite{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		company := &accountEntities.Company{
 			Name: "test",
@@ -196,7 +191,7 @@ func TestCreateCompany(t *testing.T) {
 
 		mockWrite.On("StartTransaction").Return(mockTx)
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 		body, _ := json.Marshal(company)
 		r, _ := http.NewRequest(http.MethodPost, "api/companies", bytes.NewReader(body))
 
@@ -215,7 +210,6 @@ func TestCreateCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockTx := &relational.MockWrite{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		company := struct{ Test string }{Test: "test"}
 
@@ -226,7 +220,7 @@ func TestCreateCompany(t *testing.T) {
 
 		mockWrite.On("StartTransaction").Return(mockTx)
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 		body, _ := json.Marshal(company)
 		r, _ := http.NewRequest(http.MethodPost, "api/companies", bytes.NewReader(body))
 
@@ -245,7 +239,6 @@ func TestCreateCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockTx := &relational.MockWrite{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 		appConfig := &app.Config{}
 
 		company := &accountEntities.CompanyApplicationAdmin{
@@ -265,7 +258,7 @@ func TestCreateCompany(t *testing.T) {
 			AccountID: uuid.New(),
 		}))
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, appConfig)
+		handler := NewHandler(mockWrite, mockRead, brokerMock, appConfig)
 		body, _ := json.Marshal(company)
 		r, _ := http.NewRequest(http.MethodPost, "api/companies", bytes.NewReader(body))
 
@@ -289,7 +282,6 @@ func TestCreateCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockTx := &relational.MockWrite{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 		appConfig := &app.Config{}
 
 		company := &accountEntities.CompanyApplicationAdmin{
@@ -307,7 +299,7 @@ func TestCreateCompany(t *testing.T) {
 		mockRead.On("SetFilter").Return(&gorm.DB{})
 		mockRead.On("Find").Return(response.NewResponse(0, errorsEnum.ErrNotFoundRecords, nil))
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, appConfig)
+		handler := NewHandler(mockWrite, mockRead, brokerMock, appConfig)
 		body, _ := json.Marshal(company)
 		r, _ := http.NewRequest(http.MethodPost, "api/companies", bytes.NewReader(body))
 
@@ -330,7 +322,6 @@ func TestCreateCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockTx := &relational.MockWrite{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 		appConfig := &app.Config{}
 
 		company := &accountEntities.CompanyApplicationAdmin{
@@ -348,7 +339,7 @@ func TestCreateCompany(t *testing.T) {
 		mockRead.On("SetFilter").Return(&gorm.DB{})
 		mockRead.On("Find").Return(response.NewResponse(0, errors.New("unexpected error"), nil))
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, appConfig)
+		handler := NewHandler(mockWrite, mockRead, brokerMock, appConfig)
 		body, _ := json.Marshal(company)
 		r, _ := http.NewRequest(http.MethodPost, "api/companies", bytes.NewReader(body))
 
@@ -371,7 +362,6 @@ func TestCreateCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockTx := &relational.MockWrite{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		company := &accountEntities.Company{
 			Name: "test",
@@ -384,7 +374,7 @@ func TestCreateCompany(t *testing.T) {
 
 		mockWrite.On("StartTransaction").Return(mockTx)
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 		body, _ := json.Marshal(company)
 		r, _ := http.NewRequest(http.MethodPost, "api/companies", bytes.NewReader(body))
 
@@ -409,7 +399,6 @@ func TestUpdateCompany(t *testing.T) {
 		mockRead := &relational.MockRead{}
 		mockWrite := &relational.MockWrite{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		company := &accountEntities.Company{Name: "test"}
 		resp := &response.Response{}
@@ -417,7 +406,7 @@ func TestUpdateCompany(t *testing.T) {
 		mockWrite.On("Update").Return(resp)
 		mockRead.On("SetFilter").Return(&gorm.DB{})
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 		body, _ := json.Marshal(company)
 		r, _ := http.NewRequest(http.MethodPatch, "api/companies/123", bytes.NewReader(body))
 		w := httptest.NewRecorder()
@@ -431,7 +420,6 @@ func TestUpdateCompany(t *testing.T) {
 		mockRead := &relational.MockRead{}
 		mockWrite := &relational.MockWrite{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		company := &accountEntities.Company{Name: "test"}
 		resp := &response.Response{}
@@ -439,7 +427,7 @@ func TestUpdateCompany(t *testing.T) {
 		mockWrite.On("Update").Return(resp)
 		mockRead.On("SetFilter").Return(&gorm.DB{})
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 		body, _ := json.Marshal(company)
 		r, _ := http.NewRequest(http.MethodPatch, "api/companies/123", bytes.NewReader(body))
 		w := httptest.NewRecorder()
@@ -453,9 +441,8 @@ func TestUpdateCompany(t *testing.T) {
 		mockRead := &relational.MockRead{}
 		mockWrite := &relational.MockWrite{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		r, _ := http.NewRequest(http.MethodPatch, "api/companies/123", bytes.NewReader([]byte("")))
 		w := httptest.NewRecorder()
@@ -471,7 +458,6 @@ func TestGetCompany(t *testing.T) {
 		mockRead := &relational.MockRead{}
 		mockWrite := &relational.MockWrite{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		company := &accountEntities.Company{Name: "test"}
 		resp := &response.Response{}
@@ -479,7 +465,7 @@ func TestGetCompany(t *testing.T) {
 		mockRead.On("Find").Return(resp)
 		mockRead.On("SetFilter").Return(&gorm.DB{})
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 		r, _ := http.NewRequest(http.MethodPost, "api/companies/123", nil)
 		w := httptest.NewRecorder()
 
@@ -492,14 +478,13 @@ func TestGetCompany(t *testing.T) {
 		mockRead := &relational.MockRead{}
 		mockWrite := &relational.MockWrite{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		resp := &response.Response{}
 		resp.SetError(errors.New("test"))
 		mockRead.On("Find").Return(resp)
 		mockRead.On("SetFilter").Return(&gorm.DB{})
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 		r, _ := http.NewRequest(http.MethodPost, "api/companies/123", nil)
 		w := httptest.NewRecorder()
 
@@ -571,7 +556,6 @@ func TestUpdateAccountCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		accountCompany := &roles.AccountCompany{Role: "admin", AccountID: uuid.New(), CompanyID: uuid.New()}
 		companiesResp := &response.Response{}
@@ -579,7 +563,7 @@ func TestUpdateAccountCompany(t *testing.T) {
 		mockRead.On("Find").Return(companiesResp)
 		mockRead.On("SetFilter").Return(&gorm.DB{})
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 		body, _ := json.Marshal(accountCompany)
 		r, _ := http.NewRequest(http.MethodPost, "api/companies/123", bytes.NewReader(body))
 		w := httptest.NewRecorder()
@@ -600,13 +584,12 @@ func TestUpdateAccountCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		accountCompany := &roles.AccountCompany{Role: "test", AccountID: uuid.New(), CompanyID: uuid.New()}
 		companiesResp := &response.Response{}
 		mockWrite.On("Update").Return(companiesResp.SetData(accountCompany))
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 		body, _ := json.Marshal(accountCompany)
 		r, _ := http.NewRequest(http.MethodPost, "api/companies/123", bytes.NewReader(body))
 		w := httptest.NewRecorder()
@@ -623,7 +606,6 @@ func TestUpdateAccountCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		accountCompany := &roles.AccountCompany{Role: "admin", AccountID: uuid.New(), CompanyID: uuid.New()}
 		companiesResp := &response.Response{}
@@ -631,7 +613,7 @@ func TestUpdateAccountCompany(t *testing.T) {
 		mockRead.On("Find").Return(companiesResp)
 		mockRead.On("SetFilter").Return(&gorm.DB{})
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 		body, _ := json.Marshal(accountCompany)
 		r, _ := http.NewRequest(http.MethodPost, "api/companies/123", bytes.NewReader(body))
 		w := httptest.NewRecorder()
@@ -653,9 +635,8 @@ func TestUpdateAccountCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		accountCompany := &roles.AccountCompany{Role: "admin", AccountID: uuid.New(), CompanyID: uuid.New()}
 		body, _ := json.Marshal(accountCompany)
@@ -671,9 +652,8 @@ func TestUpdateAccountCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		accountCompany := &roles.AccountCompany{Role: "admin", AccountID: uuid.New(), CompanyID: uuid.New()}
 		body, _ := json.Marshal(accountCompany)
@@ -711,7 +691,6 @@ func TestInviteUser(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		respCompany := &response.Response{}
 		respAccount := &response.Response{}
@@ -721,7 +700,7 @@ func TestInviteUser(t *testing.T) {
 		brokerMock.On("Publish").Return(nil)
 		mockWrite.On("Create").Return(respCompany)
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		body, _ := json.Marshal(inviteUser)
 		r, _ := http.NewRequest(http.MethodPost, "api/companies/", bytes.NewReader(body))
@@ -740,7 +719,6 @@ func TestInviteUser(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		respCompany := &response.Response{}
 		respAccount := &response.Response{}
@@ -750,7 +728,7 @@ func TestInviteUser(t *testing.T) {
 		brokerMock.On("Publish").Return(errors.New("test"))
 		mockWrite.On("Create").Return(respCompany)
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		body, _ := json.Marshal(inviteUser)
 		r, _ := http.NewRequest(http.MethodPost, "api/companies/", bytes.NewReader(body))
@@ -769,7 +747,6 @@ func TestInviteUser(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		respCompany := &response.Response{}
 		respAccount := &response.Response{}
@@ -781,7 +758,7 @@ func TestInviteUser(t *testing.T) {
 		mockWrite.On("Create").Return(
 			respWithError.SetError(errors.New(errorsEnum.ErrorAlreadyExistingCompanyID)))
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		body, _ := json.Marshal(inviteUser)
 		r, _ := http.NewRequest(http.MethodPost, "api/companies/", bytes.NewReader(body))
@@ -800,7 +777,6 @@ func TestInviteUser(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		respCompany := &response.Response{}
 		respAccount := &response.Response{}
@@ -811,7 +787,7 @@ func TestInviteUser(t *testing.T) {
 		mockRead.On("SetFilter").Return(&gorm.DB{})
 		mockWrite.On("Create").Return(respWithError.SetError(errorsEnum.ErrNotFoundRecords))
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		body, _ := json.Marshal(inviteUser)
 		r, _ := http.NewRequest(http.MethodPost, "api/companies/", bytes.NewReader(body))
@@ -830,9 +806,8 @@ func TestInviteUser(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		body, _ := json.Marshal(inviteUser)
 		r, _ := http.NewRequest(http.MethodPost, "api/companies/", bytes.NewReader(body))
@@ -847,9 +822,8 @@ func TestInviteUser(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		r, _ := http.NewRequest(http.MethodPost, "api/companies/", bytes.NewReader([]byte("")))
 		w := httptest.NewRecorder()
@@ -865,12 +839,11 @@ func TestDeleteCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		resp := &response.Response{}
 		mockWrite.On("Delete").Return(resp)
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		r, _ := http.NewRequest(http.MethodDelete, "api/companies/", nil)
 		w := httptest.NewRecorder()
@@ -888,12 +861,11 @@ func TestDeleteCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		resp := &response.Response{}
 		mockWrite.On("Delete").Return(resp.SetError(errors.New("test")))
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		r, _ := http.NewRequest(http.MethodDelete, "api/companies/", nil)
 		w := httptest.NewRecorder()
@@ -911,9 +883,8 @@ func TestDeleteCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		r, _ := http.NewRequest(http.MethodDelete, "api/companies/", nil)
 		w := httptest.NewRecorder()
@@ -929,14 +900,13 @@ func TestGetAccountsFromCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		accounts := &[]roles.AccountRole{{Email: "test@test.com", Username: "test", Role: "member"}}
 
 		accountsResp := &response.Response{}
 		mockRead.On("RawSQL").Return(accountsResp.SetData(accounts))
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		r, _ := http.NewRequest(http.MethodGet, "api/companies/", nil)
 		w := httptest.NewRecorder()
@@ -954,12 +924,11 @@ func TestGetAccountsFromCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		accountsResp := &response.Response{}
 		mockRead.On("RawSQL").Return(accountsResp.SetError(errors.New("test")))
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		r, _ := http.NewRequest(http.MethodGet, "api/companies/", nil)
 		w := httptest.NewRecorder()
@@ -977,9 +946,8 @@ func TestGetAccountsFromCompany(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		r, _ := http.NewRequest(http.MethodGet, "api/companies/", nil)
 		w := httptest.NewRecorder()
@@ -997,14 +965,13 @@ func TestRemoveUser(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		resp := &response.Response{}
 		mockWrite.On("Delete").Return(resp)
 		mockRead.On("Find").Return(resp.SetData(account))
 		mockRead.On("SetFilter").Return(&gorm.DB{})
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		r, _ := http.NewRequest(http.MethodDelete, "api/companies/", nil)
 		w := httptest.NewRecorder()
@@ -1023,14 +990,13 @@ func TestRemoveUser(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		resp := &response.Response{}
 		mockWrite.On("Delete").Return(resp)
 		mockRead.On("Find").Return(resp.SetError(errors.New("test")))
 		mockRead.On("SetFilter").Return(&gorm.DB{})
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		r, _ := http.NewRequest(http.MethodDelete, "api/companies/", nil)
 		w := httptest.NewRecorder()
@@ -1049,14 +1015,13 @@ func TestRemoveUser(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		resp := &response.Response{}
 		mockWrite.On("Delete").Return(resp)
 		mockRead.On("Find").Return(resp.SetError(errors.New("test")))
 		mockRead.On("SetFilter").Return(&gorm.DB{})
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		r, _ := http.NewRequest(http.MethodDelete, "api/companies/", nil)
 		w := httptest.NewRecorder()
@@ -1075,14 +1040,13 @@ func TestRemoveUser(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		resp := &response.Response{}
 		mockWrite.On("Delete").Return(resp)
 		mockRead.On("Find").Return(resp.SetError(errorsEnum.ErrNotFoundRecords))
 		mockRead.On("SetFilter").Return(&gorm.DB{})
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		r, _ := http.NewRequest(http.MethodDelete, "api/companies/", nil)
 		w := httptest.NewRecorder()
@@ -1101,14 +1065,13 @@ func TestRemoveUser(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		resp := &response.Response{}
 		mockWrite.On("Delete").Return(resp)
 		mockRead.On("Find").Return(resp.SetError(errorsEnum.ErrNotFoundRecords))
 		mockRead.On("SetFilter").Return(&gorm.DB{})
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		r, _ := http.NewRequest(http.MethodDelete, "api/companies/", nil)
 		w := httptest.NewRecorder()
@@ -1126,14 +1089,13 @@ func TestRemoveUser(t *testing.T) {
 		mockWrite := &relational.MockWrite{}
 		mockRead := &relational.MockRead{}
 		brokerMock := &broker.Mock{}
-		cacheRepositoryMock := &cache.Mock{}
 
 		resp := &response.Response{}
 		mockWrite.On("Delete").Return(resp)
 		mockRead.On("Find").Return(resp.SetError(errorsEnum.ErrNotFoundRecords))
 		mockRead.On("SetFilter").Return(&gorm.DB{})
 
-		handler := NewHandler(mockWrite, mockRead, cacheRepositoryMock, brokerMock, &app.Config{})
+		handler := NewHandler(mockWrite, mockRead, brokerMock, &app.Config{})
 
 		r, _ := http.NewRequest(http.MethodDelete, "api/companies/", nil)
 		w := httptest.NewRecorder()
